spacchetti-cli → spago
