+++ templates/gitignore
@@ -7,4 +7,4 @@
 /.psc*
 /.purs*
 /.psa*
-/.spacchetti
+/.spago
