+++ templates/spago.dhall
@@ -1,5 +1,5 @@
 {-
-Welcome to a Spacchetti project!
+Welcome to a Spago project!
 You can edit this file as you like.
 -}
 { name =
